Updated management commands to use multiple schools
