+++ django_project/core/management/commands/update_terms.py
@@ -1,0 +1,9 @@
+from ._command import MyBaseCommand
+
+
+class Command(MyBaseCommand):
+	help = 'Update term information'
+
+	def handle_school(self, scraper, options):
+
+		scraper.update_terms()
